Add files via upload
This is our team presentation documentation on a bank application enhancement. The purpose of this enhancement is to help freelancers get loans easier, as well as organize their finances. Our enhancement uses AI and financial summaries to help users and was thought of as an idea of social finance for a bank.